+++ climatoology/store/exception.py
@@ -1,0 +1,4 @@
+class InfoNotReceivedError(Exception):
+    """A plugin did not respond in time."""
+
+    pass